Functional FIFO

+++ program/FairQueue.java
@@ -3,8 +3,8 @@ import java.util.LinkedList;
 public class FairQueue implements Schedule {
 
 	/* Attributes */	
-	public double packets_dropped_size;
-	public double packets_switched_size;
+	public int packets_dropped_size;
+	public int packets_switched_size;
 	public int packets_dropped_cnt;
 	public int packets_switched_cnt;
 	public int total_wait_time;
